Refactor import paths in test, remove surplus tests
