+++ packages/location-management/src/components/LocationTree/tests/index.test.tsx
@@ -1,7 +1,7 @@
 import { mount } from 'enzyme';
 import React from 'react';
 import Tree from '../';
-import { treedata } from './fixtures';
+import { treedata } from '../../../ducks/locationHierarchy/tests/hierarchyFixtures';
 import { store } from '@onaio/redux-reducer-registry';
 import { Provider } from 'react-redux';
 import { act } from 'react-dom/test-utils';
